requirements.txt
albumentations==1.2.1
debugpy==1.6.2
diagrams==0.23.1
fvcore==0.1.5.post20220512
nbformat==5.4.0
numpy==1.21.6
omegaconf==2.2.2
opencv_python==4.6.0.66
Pillow==9.3.0
Pygments==2.12.0
pytesseract==0.3.9
PyYAML==6.0
scipy==1.7.3
Shapely==1.8.2
torch==1.12.0
tqdm==4.64.0
torchvision
blobfile
mypy
pytest
requests
einops
tensorboardX
